Removal of most unused modules, added initial Plist parser.
Removel the full service module. This includes the scl, scl-thrift,
cc-api cc-api-thrift modules.
The still needed classes were added in the plugin module under runtime.

Added new plist parser implemented with dd-plist api.

Added dummy error report, with burnt in paths for plist file and porject,
but the a valid checker-id will be parsed, and displayed on the ui.
An inclusion of a two step bugpath also included to test the jump to
bugposition feature.