Finalized project structure
Modified project structure according to a widely used template.

Replaced the old dependency copy mechanism that copied the needed jars
under the plugin/lib folder, with a more osgi compliant solution, made
possible with the p2-maven-plugin.

Added Oxygen compatible target platform. Now this is the oldest
officially supported release of this plugin. At the same time the
required lowest Eclipse-PDE should be Eclipse Oxygen 1.

Added Integration tests as well. Swtbot is being used to conduct the gui
testing. There is a dummy test added. This is needed for the test runner
to be able to function properly.