Extended project reconfiguration capabilities
Project environment variables are getting reconfigured on global/project
level changing.

Also on global configuration change the projects that set to use the
global configuration also get updated.

Added new logical representation of a CodeChecker project. All
configuration belongs to this projects, except the CcGlobalConfiguration
which is a singleton.
This new class gets the responsibility of modifying the projects
environment variables, and path management.

The CodeCheckerContext now stores these projects instead the induvidual
configurations.

Added an extra requirement to the pom of the integration tests resolve
that CDT was missing the project templates in CLI runs.