Add new CodeChecker resolution logic
Refactored the logic, how CodeChecker is found and configured.
There is now an internal representation of a CodeChecker.
There are a locator service introduced to handle CodeChecker resolution.
There are a new gui radio button group to select how CodeChecker should
be found.
There is a proper dynamic widget enablement function depending on the
selected options, and more verbose messages for the preferences panel.