Add tests for resolution logic
New Unit tests and extended Integration tests.
The CodeChecker class, and the locator logic is tested for PATH and Pre
built package resolution.
This is the same for the Integration tests, so only these two are
tested. The Custom built CodeChecker package with virtual env needs a
virtual environment emulation, which is not implemented yet.