Purged virtual environment relations
Codechecker now supports a standalone package target, resulting that a
python environment is unnecessary.

Removed from both the gui and from the internal parts.

Also there is a small change in the package binary directory widget. The
default listener was changed to a modify listener for robustness, and
for an easier guitesting.