Parameter passthrough
There is a new feature of CodeChecker configuration. Now there is a
direct parameter passthrough to the CodeChecker package. It can be
configured in the preferences page.

There is also a live display of the analyze command right below the
passthrough specification Text widget. Uneditable, but can be copied
from. With this, an actual analyze run is more easily reproducible.