Fixed hierarchical checker display (#182)
* Fixed hierarchical checker display

The old algorithm failed to add bugpaths in the following case:

Checker display set to hierarchical.
unix
    - unix.MallocSizeof
    - unix.Malloc       <- ArrayIndexOutOfBounds exception when clicked.

unix.MallocSizeof starts with unix.Malloc and was tried to be added
as a child of unix.Malloc. This resulted the aforementioned exception.

The modified algorithm adds the bugpaths correctly and should work with deeper
checker trees also.